Done semester, course and teacher

+++ src/modules/course/course.service.ts
@@ -11,6 +11,9 @@ export class CourseService {
 
   async createCourse(data) {
     try {
+      if (!(data.code.length === 6)) {
+        throw new Error('Code must be 6 characters long');
+      }
       // validate khong duoc tao trung course trong 1 ky
       const course = await this.courseRepository.findOne({
         where: {
@@ -22,7 +25,15 @@ export class CourseService {
       if (course) {
         throw new Error('You can not create duplicate course in a semester');
       }
-      const newCourse = this.courseRepository.create(data);
+      if (!await this.checkSemesterStart(data.semesterId)) {
+        throw new Error('Semester started, can not create course in this semester');
+      }
+
+      const newCourse = this.courseRepository.create({
+        ...data,
+        isActive: false
+      });
+
       return await this.courseRepository.save(newCourse);
     } catch (error) {
       throw new Error(error);
@@ -140,6 +151,11 @@ export class CourseService {
       if (!course) {
         throw new Error('Course not found');
       }
+
+      const isSemesterUsed = await this.checkSemesterIsUsed(courseId);
+      if (isSemesterUsed && data.isActive === false) {
+        throw new Error('Course is used in class');
+      }
       return await this.courseRepository.save({ ...course, ...data });
     } catch (error) {
       throw new Error(error);
@@ -152,9 +168,35 @@ export class CourseService {
       if (!course) {
         throw new Error('Course not found');
       }
+      if (course.isActive) {
+        throw new Error('Course is in used');
+      }
       return await this.courseRepository.remove(course);
     } catch (error) {
       throw new Error(error);
     }
   }
+
+  async checkSemesterIsUsed(courseId: string) {
+    // kiểm tra xem semester có được sử dụng trong course không
+    const course = await this.dataSource
+      .getRepository('class')
+      .createQueryBuilder('class')
+      .where('class.courseId = :courseId', { courseId })
+      .getOne();
+
+    return course ? true : false;
+  }
+
+  //check semester have end less than current date
+  async checkSemesterStart(semesterId: string): Promise<boolean> {
+    const semester = await this.dataSource
+      .getRepository('semester')
+      .createQueryBuilder('semester')
+      .where('semester.id = :semesterId', { semesterId })
+      .andWhere('semester.startDate > :currentDate', { currentDate: new Date() })
+      .getOne();
+  
+    return semester ? true : false;
+  }  
 }
